Notebook.cs - Moved the UpdateLocalStorage(this) in the Save method until after it tries to save the Notebook in the database, (so it uses the correct id)
Added a rootcontextmenu (you can only add folders to the 'My Notebooks' treeviewitem).
Simplified contextmenucreation, and moved it to the OnActivation function (instead of constructor)
Moved the loading of the complete treeview (containing, folders and notebooks) to its own method.
Changed the methods LoadFolders and LoadNotebooks to work with List<TreeViewItem> instead of TreeView
Added AddFolderToRoot function (because getting the path is different for root)
Load NotebookLocation instead of locationID (in AddNotebook)
Changed GetPath() to work with new setup (of TreeView)